# docs/specs/openapi.yaml
openapi: "3.1.0"
info:
  title: Aether Platform API
  description: |
    White-label SaaS коммуникационная платформа.
    Multi-tenant, plugin architecture, AI-powered.

    ## Аутентификация
    - **Bearer JWT**: `Authorization: Bearer <access_token>`
    - **API Key**: `X-API-Key: ak_...`

    ## Tenant isolation
    Все tenant-specific эндпоинты требуют JWT с `organisation_id` claim.
    Admin эндпоинты требуют `is_superadmin = true`.

    ## Rate Limiting
    - Auth: IP-based (5 login/min, 3 signup/hour)
    - Tenant: per-plan (60-6000 RPM)
    - Headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`

    ## Pagination
    Cursor-based: `?cursor=<last_id>&limit=<1-100>`, default 20.
    Response: `{"items": [...], "next_cursor": "uuid"}`.

    ## Errors
    Все ошибки возвращают JSON: `{"error": {"code": "...", "message": "...", "details": {...}}}`
  version: "1.0.0"
  contact:
    name: Aether Platform
    email: dev@aether.cloud
servers:
  - url: https://api.aether.cloud/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Development
security:
  - bearerAuth: []
  - apiKeyAuth: []

# ============================================================================
# SECURITY SCHEMES
# ============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login or /auth/refresh
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 'API key from /auth/api-keys (prefix: ak_)'

  # ============================================================================
  # COMMON SCHEMAS
  # ============================================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "invalid_credentials"
            message:
              type: string
              description: Human-readable error message (localized)
              example: "Неверный email или пароль"
            details:
              type: object
              description: Field-level validation errors
...
    PaginatedResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {}
        next_cursor:
          type: string
          format: uuid
          nullable: true
          description: Cursor for next page, null if last page
        total:
          type: integer
          description: Total count (when available)
          example: 42

    # --- Auth ---
...
    InvoiceResponse:
      type: object
      required: [id, amount_usd, currency, status, due_date, created_at]
      properties:
        id:
          type: string
          format: uuid
        amount_usd:
          type: number
          format: float
        currency:
          type: string
        status:
          type: string
          enum: [draft, open, paid, void, past_due]
        due_date:
          type: string
          format: date
        paid_at:
          type: string
          format: date-time
          nullable: true
        invoice_pdf_url:
          type: string
          format: uri
          nullable: true
        created_at:
          type: string
          format: date-time
    UsageResponse:
      type: object
      properties:
        api_calls:
          type: object
          properties:
            current: { type: integer }
            limit: { type: integer }
        tokens_used:
          type: object
          properties:
            current: { type: integer }
            limit: { type: integer }
        storage_bytes:
          type: object
          properties:
            current: { type: integer }
            limit: { type: integer }
        active_users:
          type: object
          properties:
            current: { type: integer }
            limit: { type: integer }
        conversations_today:
          type: object
          properties:
            current: { type: integer }
            limit: { type: integer }

    # --- Tenant/Admin ---
...
  parameters:
    CursorParam:
      name: cursor
      in: query
      schema:
        type: string
        format: uuid
      description: Cursor для пагинации (id последнего элемента предыдущей страницы)
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Количество элементов на странице

  # ============================================================================
  # COMMON RESPONSES
  # ============================================================================
  responses:
    401Unauthorized:
      description: Не авторизован (JWT истёк или отсутствует)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "unauthorized"
              message: "Требуется аутентификация"
...
    404NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /billing/invoices:
    get:
      operationId: listInvoices
      summary: Счета
      tags: [Billing]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Список счетов
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/InvoiceResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
  /billing/invoices/{invoice_id}:
    get:
      operationId: getInvoice
      summary: Детали счёта
      tags: [Billing]
      security:
        - bearerAuth: []
      parameters:
        - name: invoice_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Детали счёта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
...
  /billing/usage:
    get:
      operationId: getUsage
      summary: Статистика использования
      tags: [Billing]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Текущее использование
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'

  # ==========================================================================
  # TENANTS (user-scoped)
  # ==========================================================================
...